started refactoring simclr to include decoupled contrastive learning as well as data parallelism as parameters
