+++ patchwork/tests/test_feature_contrastive.py
@@ -1,0 +1,1 @@
+import numpy as npimport tensorflow as tffrom patchwork.feature._contrastive import _build_negative_maskfrom patchwork.feature._contrastive import _simclr_softmax_prob, _hcl_softmax_probdef test_build_negative_mask():    gbs = 7    mask = _build_negative_mask(gbs)    assert mask.shape == (2*gbs, 2*gbs)    assert (mask.sum(0) == 2*gbs - 2).all()def test_simclr_softmax_prob_trivial_case():    gbs = 7    d = 23    mask = _build_negative_mask(gbs)    z = tf.nn.l2_normalize(np.random.normal(0,1, (gbs,d)))    softmax_prob, nce_batch_acc = _simclr_softmax_prob(z,z,1,mask)        assert softmax_prob.shape == (2*gbs,)    assert (softmax_prob.numpy() >= 0).all()    assert (softmax_prob.numpy() <= 1).all()    assert nce_batch_acc.shape == ()    assert nce_batch_acc.numpy() > 0.9    def test_simclr_softmax_prob_random_case():    gbs = 7    d = 23    mask = _build_negative_mask(gbs)    z1 = tf.nn.l2_normalize(np.random.normal(0,1, (gbs,d)))    z2 = tf.nn.l2_normalize(np.random.normal(0,1, (gbs,d)))    softmax_prob, nce_batch_acc = _simclr_softmax_prob(z1,z2,1,mask)        assert softmax_prob.shape == (2*gbs,)    assert (softmax_prob.numpy() >= 0).all()    assert (softmax_prob.numpy() <= 1).all()    assert nce_batch_acc.shape == ()    assert nce_batch_acc.numpy() < 0.5            def test_hcl_softmax_prob_trivial_case():    gbs = 7    d = 23    mask = _build_negative_mask(gbs)    z = tf.nn.l2_normalize(np.random.normal(0,1, (gbs,d)).astype(np.float32))    softmax_prob, nce_batch_acc = _hcl_softmax_prob(z,z,1,1,0.1,mask)        assert softmax_prob.shape == (2*gbs,)    assert (softmax_prob.numpy() >= 0).all()    assert (softmax_prob.numpy() <= 1).all()    assert nce_batch_acc.shape == ()    assert nce_batch_acc.numpy() > 0.9        def test_hcl_softmax_prob_random_case():    gbs = 7    d = 23    mask = _build_negative_mask(gbs)    z1 = tf.nn.l2_normalize(np.random.normal(0,1, (gbs,d)).astype(np.float32))    z2 = tf.nn.l2_normalize(np.random.normal(0,1, (gbs,d)).astype(np.float32))    softmax_prob, nce_batch_acc = _hcl_softmax_prob(z1,z2,1,1,0.1,mask)        assert softmax_prob.shape == (2*gbs,)    assert (softmax_prob.numpy() >= 0).all()    assert (softmax_prob.numpy() <= 1).all()    assert nce_batch_acc.shape == ()    assert nce_batch_acc.numpy() < 0.5
